Split classes into different files
