+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00f8480db01347a0a39bf4bab23e3303
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
