+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/CommandHandlers.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1f4ddfc171d7435a9890cc009f66bdb2
+timeCreated: 1536998039
