+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/NoComponentsNeeded.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 76e332bc95644fa595389f57896df709
+timeCreated: 1536998078
