+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/Readers.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6f47557fa5794ce6ad185b176cbaac35
+timeCreated: 1536998089
