+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/ReadersAndWriters.cs
@@ -1,0 +1,19 @@
+using Generated.Improbable;
+using NUnit.Framework;
+using UnityEngine;
+
+namespace Improbable.Gdk.GameObjectRepresentation.EditModeTests.MonoBehaviourActivationManagerTests
+{
+    [TestFixture]
+    public class ReadersAndWriters :
+        TestForBehaviourThatNeedsAuthority<ReadersAndWriters.TestBehaviourWithReaderAndWriter>
+    {
+        protected override uint ComponentId => new Position.Component().ComponentId;
+
+        public class TestBehaviourWithReaderAndWriter : MonoBehaviour
+        {
+            [Require] private Position.Requirable.Reader positionReader;
+            [Require] private Position.Requirable.Writer positionWriter;
+        }
+    }
+}
