+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/ReadersAndWriters.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 071ffc215f924f1f9b11bd04a3fb2d0a
+timeCreated: 1536998167
