+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/TestForBehaviourThatNeedsAuthority.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: bc5e04a9ef414aab8c99cc48645e796e
+timeCreated: 1536998096
