+++ test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManager/Writers.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cea635a5dfdd4b928b25e8a8064d43ca
+timeCreated: 1536998105
