--- test-project/Assets/EditmodeTests/GameObjectRepresentation/MonoBehaviourActivationManagerTests.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6b8e7e9e7d77439e9f7a7d6f3c10fb76
-timeCreated: 1536832556
